third_party/blink/web_tests/external/wpt/dom/events/webkit-animation-start-event-expected.txt

This is a testharness.js-based test.
FAIL onanimationstart and onwebkitanimationstart are not aliases assert_equals: onwebkitanimationstart should initially be null expected (object) null but got (undefined) undefined
FAIL dispatchEvent of a webkitAnimationStart event does trigger a prefixed event handler or listener assert_equals: onwebkitanimationstart must be supported and not previously set expected (object) null but got (undefined) undefined
PASS dispatchEvent of a webkitAnimationStart event does not trigger an unprefixed event handler or listener
FAIL dispatchEvent of an animationstart event does not trigger a prefixed event handler or listener assert_equals: onwebkitanimationstart must be supported and not previously set expected (object) null but got (undefined) undefined
FAIL onwebkitanimationstart event handler should trigger for an animation assert_equals: onwebkitanimationstart must be supported and not previously set expected (object) null but got (undefined) undefined
FAIL onwebkitanimationstart event handler should not trigger if an unprefixed event handler also exists assert_equals: onwebkitanimationstart must be supported and not previously set expected (object) null but got (undefined) undefined
FAIL onwebkitanimationstart event handler should not trigger if an unprefixed listener also exists assert_equals: onwebkitanimationstart must be supported and not previously set expected (object) null but got (undefined) undefined
FAIL event types for prefixed and unprefixed animationstart event handlers should be named appropriately assert_equals: onwebkitanimationstart must be supported and not previously set expected (object) null but got (undefined) undefined
PASS webkitAnimationStart event listener should trigger for an animation
PASS webkitAnimationStart event listener should not trigger if an unprefixed listener also exists
PASS webkitAnimationStart event listener should not trigger if an unprefixed event handler also exists
PASS event types for prefixed and unprefixed animationstart event listeners should be named appropriately
PASS webkitAnimationStart event listener is case sensitive
Harness: the test ran to completion.
